% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{akkadianRoot}
\alias{akkadianRoot}
\title{The inferred method of Akkadian-Babylonian mathematicians to approximate
square roots}
\usage{
akkadianRoot(a, n = 100L)
}
\arguments{
\item{a}{the number to compute square root of}

\item{n}{the number of iterations to take in the method}
}
\description{
{For finding the square-root of whole numbers.}
}
\details{
{The algorithm is on wikipedia and quite straightforward.}
}
