% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uniform_log_util}
\alias{uniform_log_util}
\title{Newton-Raphson algorithm for solving optimal log utiltiy under uniform distributions}
\usage{
uniform_log_util(interval, rate = 0, n = 500L)
}
\arguments{
\item{interval}{the interval for the uniform distribution}

\item{rate}{the risk-neutral rate}

\item{n}{the number of iterations to take in the NR method.}
}
\description{
{For finding the root of the function \eqn{h(x)=\log((1+bx)/(1+ax))/(b-a)-x}.}
}
\details{
{The algorithm is on wikipedia and quite straightforward.}
}
