% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrators.R
\name{multidim_integrator}
\alias{multidim_integrator}
\title{Multi-dimensional integration via Monte-Carlo methods}
\usage{
multidim_integrator(g, lbs, ubs, n = 10^4, sig_lvl = 0.05, ...)
}
\arguments{
\item{g}{the known integrand, a function of a vector defined over a 'rectangular' region}

\item{lbs}{the left-end points of each interval per coordinate}

\item{ubs}{the right end-points of each interval per coordinate}

\item{n}{the number of variates to simulate of the IID uniform vector}

\item{sig_lvl}{the significance level for the confidence intervals}

\item{...}{additional arguments to pass to the function \code{g}}
}
\value{
data.frame containing
\itemize{
\item \code{estimate} the point estimate of the integral,
\item \code{lb} the lower bound of the confidence interval,
\item \code{ub} the upper bound of the confidence interval,
\item \code{std_error} the standard error of the point-estimate. }
}
\description{
{A naive but efficient estimator for multi-dimensional integrals
via Monte-Carlo methods. The integrand must take a vector as an argument.}
}
